Fix Issue Linking Tweets
For a reply to be associated with the original tweet, it has to have 
in_reply_to_status_id set. For that to be set, the room has to be set.
For the room to be properly set in Lita, it must be either a Room
object or a string.

This commit casts tweet.id as a string so it will be properly
recognized as a room.